Renamed stateful to pforward
